+++ VectorEditor.Core/Structures/Rectangle.cs
@@ -176,7 +176,7 @@ public class Rectangle : IShape
     public double GetMaxY() => Math.Max(_startPoint.Y, _oppositePoint.Y);
 
     // --- GEOMETRIA (Również pilnuje blokady) ---
-    public void Move(int dx, int dy)
+    public void Move(double dx, double dy)
     {
         if (IsBlocked) return;
         _startPoint = new Point(_startPoint.X + dx, _startPoint.Y + dy);
